Customizable fingerprint maximum fail attempts [1/2]
Add an option to customize maximum fingerprint failed attempts before
showing "Too many attempts". The change requires a reboot, we may need
to add a form of listener in FingerprintService for it to work without
rebooting.

This is the frameworks/base part of the commit, another commit is in
Settings.